fix(lock): improve redis lock algo

+++ apps/backend/src/util/redis/lock.ts
@@ -1,9 +1,8 @@
-import { setTimeout } from "node:timers/promises";
 import type { createClient } from "redis";
 
 type RedisClient = ReturnType<typeof createClient>;
 
-const acquireLock = async ({
+const acquireLock = ({
   client,
   name,
   timeout,
@@ -14,45 +13,60 @@ const acquireLock = async ({
   timeout: number;
   retryDelay: number;
 }) => {
-  const result = await client.set(name, "1", {
-    PX: timeout,
-    NX: true,
-  });
+  return new Promise<string>((resolve, reject) => {
+    function tryAcquire() {
+      const rdn = Math.random().toString(36);
+      client
+        .set(name, rdn, {
+          PX: timeout,
+          NX: true,
+        })
+        .then((result) => {
+          if (result === "OK") {
+            resolve(rdn);
+          } else {
+            const adjustedTimeout = retryDelay + Math.ceil(Math.random() * 10);
+            setTimeout(tryAcquire, adjustedTimeout);
+          }
+        })
+        .catch(reject);
+    }
 
-  if (result !== "OK") {
-    await setTimeout(retryDelay);
-    await acquireLock({ client, name, timeout, retryDelay });
-  }
+    tryAcquire();
+  });
 };
 
-type Acquire = <T>(
-  name: string,
-  task: () => Promise<T>,
-  options?: { timeout?: number; retryDelay?: number },
-) => Promise<T>;
-
-interface RedisLock {
-  acquire: Acquire;
-}
-
-export const createRedisLock = (client: RedisClient): RedisLock => {
-  const acquire: Acquire = async (
-    name,
-    task,
+export const createRedisLock = (client: RedisClient) => {
+  async function acquire<T>(
+    name: string,
+    task: () => Promise<T>,
     { timeout = 20000, retryDelay = 500 } = {},
-  ) => {
+  ) {
     const fullName = `lock.${name}`;
-
-    await acquireLock({
+    const id = await acquireLock({
       client,
       name: fullName,
       timeout,
       retryDelay,
     });
-    const result = await task();
-    await client.del(fullName);
+    let timer: NodeJS.Timeout | null = null;
+    const result = (await Promise.race([
+      task(),
+      new Promise((_resolve, reject) => {
+        timer = setTimeout(() => {
+          reject(new Error("Lock timeout"));
+        }, timeout);
+      }),
+    ])) as T;
+    if (timer) {
+      clearTimeout(timer);
+    }
+    const value = await client.get(fullName);
+    if (value === id) {
+      await client.del(fullName);
+    }
     return result;
-  };
+  }
 
   return { acquire };
 };
